gcp_functions/topic_assignment_daily/requirements.txt
functions-framework==3.*
pandas==2.1.0
google-cloud-bigquery==3.12.0
db-dtypes==1.1.1
google-cloud == 0.34.0
pandas-gbq == 0.22.0
stanza == 1.8.1